Open Weather Dashboard
This project involves developing an **API signature** for **OpenWeather** using the **R programming language**. The code is deployed into a web-based dashboard that presents essential weather information to the user. This dashboard is designed to help a human operator determine when to **activate or deactivate an irrigation controller**, specifically when the target soil moisture level is reached. A prospective buyer could integrate this development into a machine learning pipeline. This additional process would automate the switching of the controller using established predictive models, such as those based on **Fourier or Gaussian switching equations**. The ultimate goal is to enhance precision and efficiency in irrigation management.